+++ src/fides/api/ops/models/storage.py
@@ -20,7 +20,7 @@ from fides.api.ops.schemas.storage.storage_secrets_docs_only import (
     possible_storage_secrets,
 )
 from fides.api.ops.util.logger import Pii
-from fides.ctl.core.config import get_config
+from fides.core.config import get_config
 from fides.lib.db.base import Base  # type: ignore[attr-defined]
 
 CONFIG = get_config()