+++ src/fides/api/ops/task/graph_task.py
@@ -43,7 +43,7 @@ from fides.api.ops.util.cache import get_cache
 from fides.api.ops.util.collection_util import NodeInput, Row, append, partition
 from fides.api.ops.util.logger import Pii
 from fides.api.ops.util.saas_util import FIDESOPS_GROUPED_INPUTS
-from fides.ctl.core.config import get_config
+from fides.core.config import get_config
 
 dask.config.set(scheduler="threads")