+++ src/fides/core/config/credentials_settings.py
@@ -4,7 +4,7 @@
 import os
 from typing import Dict, Optional
 
-from fides.ctl.connectors.models import (
+from fides.connectors.models import (
     AWSConfig,
     BigQueryConfig,
     DatabaseConfig,