+++ src/fides/core/parse.py
@@ -3,7 +3,7 @@ from fideslang import Taxonomy
 from fideslang.manifests import ingest_manifests
 from fideslang.parse import load_manifests_into_taxonomy
 
-from fides.ctl.core.utils import echo_green, get_manifest_list
+from fides.core.utils import echo_green, get_manifest_list
 
 
 def parse(manifests_dir: str) -> Taxonomy: