+++ tests/ops/service/messaging/message_dispatch_service_test.py
@@ -20,7 +20,7 @@ from fides.api.ops.schemas.messaging.messaging import (
 )
 from fides.api.ops.schemas.redis_cache import Identity
 from fides.api.ops.service.messaging.message_dispatch_service import dispatch_message
-from fides.ctl.core.config import get_config
+from fides.core.config import get_config
 
 CONFIG = get_config()